+++ src/test/java/org/scion/jpan/ppl/PolicyTest.java
@@ -22,14 +22,13 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.scion.jpan.Path;
 
-@SuppressWarnings("deprecation")
 class PolicyTest {
 
   private static final String ALLOW_STR = "+";
   private static final String DENY_STR = "-";
 
   @Test
-  void testBasicPolicy() {
+  void testBasicFilter() {
     PathProvider pp = new PathProvider();
     List<Path> paths = pp.getPaths("2-ff00:0:212", "2-ff00:0:211");
 
@@ -49,17 +48,24 @@ class PolicyTest {
   }
 
   @Test
+  void testBasicPolicy() {
+    // "Empty policy"
+    // Unlike the scionproto, we forbid empty policies.
+    assertThrows(PplException.class, () -> PplPolicy.builder().build());
+  }
+
+  @Test
   void testOptionsEval() {
     PathProvider pp = new PathProvider();
-    PplPathFilter policy;
+    PplPolicy policy;
     List<Path> paths;
 
     // "one option, allow everything"
     policy =
-        PplPathFilter.builder()
-            .addOption(
-                0,
-                PplExtPolicy.builder()
+        PplPolicy.builder()
+            .add(
+                "0",
+                PplPathFilter.builder()
                     .addAclEntry(true, "0-0#0")
                     .addAclEntry(DENY_STR)
                     .buildNoValidate())
@@ -69,16 +75,16 @@ class PolicyTest {
 
     // "two options, deny everything"
     policy =
-        PplPathFilter.builder()
-            .addOption(
-                0,
-                PplExtPolicy.builder()
+        PplPolicy.builder()
+            .add(
+                "0",
+                PplPathFilter.builder()
                     .addAclEntry(true, "0-0#0")
                     .addAclEntry(DENY_STR)
                     .buildNoValidate())
-            .addOption(
-                1,
-                PplExtPolicy.builder()
+            .add(
+                "0",
+                PplPathFilter.builder()
                     .addAclEntry(false, "0-0#0")
                     .addAclEntry(DENY_STR)
                     .buildNoValidate())
@@ -88,17 +94,16 @@ class PolicyTest {
 
     // "two options, first: allow everything, second: allow one path"
     policy =
-        PplPathFilter.builder()
-            .setName("")
-            .addOption(
-                0,
-                PplExtPolicy.builder()
+        PplPolicy.builder()
+            .add(
+                "5",
+                PplPathFilter.builder()
                     .addAclEntry(true, "0-0#0")
                     .addAclEntry(DENY_STR)
                     .buildNoValidate())
-            .addOption(
-                1,
-                PplExtPolicy.builder()
+            .add(
+                "0",
+                PplPathFilter.builder()
                     .addAclEntry(false, "1-ff00:0:110#0")
                     .addAclEntry(false, "1-ff00:0:120#0")
                     .addAclEntry(false, "1-ff00:0:111#2823")
@@ -108,37 +113,45 @@ class PolicyTest {
     paths = pp.getPaths("1-ff00:0:122", "2-ff00:0:222");
     assertEquals(1, policy.filter(paths).size());
 
+    // TODO this does currently not work. Do we want this?
+    //   We could allow a _list_ of filters per destination. But how are they applied?
+    //   - Allow only paths that pass all filters?
+    //   - Allow all paths that pass at least one filter?
+    //   - Look at what scionproto claims: try all filters until it finds one that matches...
+    //     In practice it seems to allow any path that passes at least one filter.
+    //   - We could also provide a more complex logic that allow combining filters with && and ||.
     // "two options, combined"
-    policy =
-        PplPathFilter.builder()
-            .addOption(
-                0,
-                PplExtPolicy.builder()
-                    .addAclEntry(false, "1-ff00:0:120#0")
-                    .addAclEntries(ALLOW_STR)
-                    .build())
-            .addOption(
-                0,
-                PplExtPolicy.builder()
-                    .addAclEntry(false, "2-ff00:0:210#0")
-                    .addAclEntry(ALLOW_STR)
-                    .build())
-            .build();
-    paths = pp.getPaths("1-ff00:0:110", "2-ff00:0:220");
-    assertEquals(3, policy.filter(paths).size());
+    //    policy =
+    //        PplPolicy.builder()
+    //            .add(
+    //                "0",
+    //                PplPathFilter.builder()
+    //                    .addAclEntry(false, "1-ff00:0:120#0")
+    //                    .addAclEntries(ALLOW_STR)
+    //                    .build(),
+    //                PplPathFilter.builder()
+    //                    .addAclEntry(false, "2-ff00:0:210#0")
+    //                    .addAclEntry(ALLOW_STR)
+    //                    .build())
+    //            .build();
+    //    paths = pp.getPaths("1-ff00:0:110", "2-ff00:0:220");
+    //    for (Path path : paths) {
+    //      System.out.println("path: " + ScionUtil.toStringPath(path.getMetadata()));
+    //    }
+    //    assertEquals(3, policy.filter(paths).size());
 
     // "two options, take first"
     policy =
-        PplPathFilter.builder()
-            .addOption(
-                1,
-                PplExtPolicy.builder()
+        PplPolicy.builder()
+            .add(
+                "0",
+                PplPathFilter.builder()
                     .addAclEntry(false, "1-ff00:0:120#0")
                     .addAclEntries(ALLOW_STR)
                     .build())
-            .addOption(
-                0,
-                PplExtPolicy.builder()
+            .add(
+                "0",
+                PplPathFilter.builder()
                     .addAclEntry(false, "2-ff00:0:210#0")
                     .addAclEntry(ALLOW_STR)
                     .build())
@@ -148,229 +161,22 @@ class PolicyTest {
 
     // "two options, take second"
     policy =
-        PplPathFilter.builder()
-            .addOption(
-                1,
-                PplExtPolicy.builder()
+        PplPolicy.builder()
+            .add(
+                "5",
+                PplPathFilter.builder()
                     .addAclEntry(false, "1-ff00:0:120#0")
                     .addAclEntry(ALLOW_STR)
                     .build())
-            .addOption(
-                10,
-                PplExtPolicy.builder()
+            .add(
+                "2",
+                PplPathFilter.builder()
                     .addAclEntry(false, "2-ff00:0:210#0")
                     .addAclEntry(ALLOW_STR)
                     .build())
             .build();
     paths = pp.getPaths("1-ff00:0:110", "2-ff00:0:220");
     assertEquals(2, policy.filter(paths).size());
-  }
-
-  @Test
-  void TestExtends() {
-    PplExtPolicy policy;
-    PplExtPolicy[] extended;
-    PplPathFilter extendedPolicy;
-    PplPathFilter pol;
-
-    // "one extends, use sub acl"
-    policy = PplExtPolicy.builder().addExtension("policy1").buildNoValidate();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder()
-              .setName("policy1")
-              .addAclEntry(true, "0-0#0")
-              .addAclEntry(DENY_STR)
-              .buildNoValidate()
-        };
-    extendedPolicy =
-        PplPathFilter.builder().addAclEntry(true, "0-0#0").addAclEntry(DENY_STR).buildNoValidate();
-
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "use option of extended policy"
-    policy = PplExtPolicy.builder().addExtension("policy1").buildNoValidate();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder()
-              .setName("policy1")
-              .addOption(
-                  1,
-                  PplExtPolicy.builder()
-                      .addAclEntry(true, "0-0#0")
-                      .addAclEntry(DENY_STR)
-                      .buildNoValidate())
-              .buildNoValidate()
-        };
-    extendedPolicy =
-        PplPathFilter.builder()
-            .addOption(
-                1,
-                PplExtPolicy.builder()
-                    .addAclEntry(true, "0-0#0")
-                    .addAclEntry(DENY_STR)
-                    .buildNoValidate())
-            .build();
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "two extends, use sub acl and list"
-    policy = PplExtPolicy.builder().addExtension("policy1").buildNoValidate();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder()
-              .setName("policy1")
-              .addAclEntry(true, "0-0#0")
-              .addAclEntry(DENY_STR)
-              .setSequence("1-ff00:0:133#1019 1-ff00:0:132#1910")
-              .buildNoValidate()
-        };
-    extendedPolicy =
-        PplPathFilter.builder()
-            .addAclEntry(true, "0-0#0")
-            .addAclEntry(DENY_STR)
-            .setSequence("1-ff00:0:133#1019 1-ff00:0:132#1910")
-            .buildNoValidate();
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "two extends, only use acl"
-    policy =
-        PplExtPolicy.builder()
-            .setSequence("1-ff00:0:133#0 1-ff00:0:132#0")
-            .addExtension("policy2")
-            .buildNoValidate();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder()
-              .setName("policy2")
-              .addAclEntry(true, "0-0#0")
-              .addAclEntry(DENY_STR)
-              .buildNoValidate(),
-          PplExtPolicy.builder()
-              .setName("policy1")
-              .setSequence("1-ff00:0:133#1019 1-ff00:0:132#1910")
-              .buildNoValidate()
-        };
-    extendedPolicy =
-        PplPathFilter.builder()
-            .addAclEntry(true, "0-0#0")
-            .addAclEntry(DENY_STR)
-            .setSequence("1-ff00:0:133#0 1-ff00:0:132#0")
-            .buildNoValidate();
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "three extends, use last list"
-    policy = PplExtPolicy.builder().addExtensions("p1", "p2", "p3").build();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder()
-              .setName("p1")
-              .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
-              .build(),
-          PplExtPolicy.builder()
-              .setName("p2")
-              .setSequence("1-ff00:0:133#1012 1-ff00:0:132#1912")
-              .build(),
-          PplExtPolicy.builder()
-              .setName("p3")
-              .setSequence("1-ff00:0:133#1013 1-ff00:0:132#1913")
-              .build()
-        };
-    extendedPolicy =
-        PplPathFilter.builder().setSequence("1-ff00:0:133#1013 1-ff00:0:132#1913").build();
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "nested extends"
-    policy = PplExtPolicy.builder().addExtension("policy1").buildNoValidate();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder().setName("policy1").addExtension("policy2").buildNoValidate(),
-          PplExtPolicy.builder().setName("policy2").addExtension("policy3").buildNoValidate(),
-          PplExtPolicy.builder()
-              .setName("policy3")
-              .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
-              .buildNoValidate()
-        };
-    extendedPolicy =
-        PplPathFilter.builder().setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911").build();
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "nested extends, evaluating order"
-    policy = PplExtPolicy.builder().addExtension("policy3").build();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder()
-              .setName("policy3")
-              .setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910")
-              .build(),
-          PplExtPolicy.builder().addExtension("policy2").build(),
-          PplExtPolicy.builder().setName("policy2").addExtensions("policy1").build(),
-          PplExtPolicy.builder()
-              .setName("policy1")
-              .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
-              .build()
-        };
-    extendedPolicy =
-        PplPathFilter.builder().setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910").build();
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "different nested extends, evaluating order"
-    policy = PplExtPolicy.builder().addExtension("policy6").build();
-    extended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder()
-              .setName("policy3")
-              .setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910")
-              .build(),
-          PplExtPolicy.builder().addExtension("policy2").build(),
-          PplExtPolicy.builder().setName("policy2").addExtension("policy1").build(),
-          PplExtPolicy.builder().setName("policy6").addExtension("policy3").build(),
-          PplExtPolicy.builder()
-              .setName("policy1")
-              .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
-              .build()
-        };
-    extendedPolicy =
-        PplPathFilter.builder().setSequence("1-ff00:0:133#1010 1-ff00:0:132#1910").build();
-    pol = PplPathFilter.policyFromExtPolicy(policy, extended);
-    assertEquals(extendedPolicy, pol);
-
-    // "TestPolicy Extend not found"
-    PplExtPolicy errPolicy = PplExtPolicy.builder().addExtension("policy1").build();
-    PplExtPolicy[] errExtended =
-        new PplExtPolicy[] {
-          PplExtPolicy.builder().setName("policy1").addExtension("policy16").build(),
-          PplExtPolicy.builder().setName("policy2").addExtension("policy3").build(),
-          PplExtPolicy.builder()
-              .setName("policy3")
-              .setSequence("1-ff00:0:133#1011 1-ff00:0:132#1911")
-              .build()
-        };
-    assertThrows(
-        PplException.class, () -> PplPathFilter.policyFromExtPolicy(errPolicy, errExtended));
-  }
-
-  @Test
-  void TestFilterOpt() {
-    PathProvider pp = new PathProvider();
-    List<Path> paths = pp.getPaths("1-ff00:0:110", "2-ff00:0:220");
-
-    // "sequence in options is ignored"
-    PplExtPolicy optPolicy = PplExtPolicy.builder().setSequence("0+ 1-ff00:0:111 0+").build();
-    PplPathFilter policy = PplPathFilter.builder().addOption(0, optPolicy).build();
-    List<Path> outPaths = policy.filterOpt(paths, PplPathFilter.FilterOptions.create(true));
-    assertEquals(3, outPaths.size());
-
-    // "sequence is ignored"
-    policy = PplPathFilter.builder().setSequence("0+ 1-ff00:0:111 0+").build();
-    outPaths = policy.filterOpt(paths, PplPathFilter.FilterOptions.create(true));
-    assertEquals(3, outPaths.size());
   }
 
   @Test